+++ floor-map-service/src/main/java/com/jake/waitlistservice/jpa/repository/FloorMapRepository.java
@@ -9,7 +9,7 @@ import java.math.BigInteger;
 import java.util.List;
 
 public interface FloorMapRepository extends JpaRepository<FloorMap, BigInteger> {
-    List<FloorMap> findAllByBusinessId(BigInteger businessId, Sort sortedBy);
+    List<FloorMap> findAllByBusinessId(long businessId, Sort sortedBy);
 
-    List<FloorMap> findAllByBusinessId(BigInteger businessId);
+    List<FloorMap> findAllByBusinessId(long businessId);
 }
